:sparkles: Add Azure AD OIDC — [#1861]
