+++ src/oidc_generics/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+from django.utils.translation import gettext_lazy as _
+
+
+class OIDCAppConfig(AppConfig):
+    name = "oidc_generics"
+    verbose_name = _("OpenID Connect")
