+++ src/oidc_generics/azure_ad_settings.py
@@ -1,0 +1,2 @@
+AZURE_AD_CUSTOM_OIDC_DB_PREFIX = "azure_ad_oidc"
+OIDC_AUTHENTICATION_CALLBACK_URL = "azure_ad_oidc:callback"
