+++ src/openforms/authentication/contrib/oidc/constants.py
@@ -4,3 +4,4 @@ DIGID_MACHTIGEN_OIDC_AUTH_SESSION_KEY = "digid_machtigen_oidc:machtigen"
 EHERKENNING_BEWINDVOERING_OIDC_AUTH_SESSION_KEY = (
     "eherkenning_bewindvoering_oidc:machtigen"
 )
+AZURE_AD_OIDC_AUTH_SESSION_KEY = "azure_ad_oidc:id"
