+++ src/openforms/authentication/contrib/oidc/tests/azure_ad/test_auth_procedure.py
@@ -1,0 +1,8 @@
+from django.test import TestCase
+
+
+class AzureADOIDCTests(TestCase):
+    @classmethod
+    def setUpTestData(cls):
+        # TODO
+        ...
