+++ src/openforms/authentication/contrib/oidc/tests/digid/test_auth_procedure.py
@@ -7,11 +7,7 @@ import requests_mock
 from furl import furl
 from rest_framework import status
 
-from digid_eherkenning_oidc_generics.models import (
-    OpenIDConnectDigiDMachtigenConfig,
-    OpenIDConnectEHerkenningBewindvoeringConfig,
-    OpenIDConnectPublicConfig,
-)
+from oidc_generics.models import OpenIDConnectPublicConfig
 from openforms.authentication.views import BACKEND_OUTAGE_RESPONSE_PARAMETER
 from openforms.forms.tests.factories import FormFactory
 
@@ -20,7 +16,7 @@ default_config = dict(
     oidc_rp_client_id="testclient",
     oidc_rp_client_secret="secret",
     oidc_rp_sign_algo="RS256",
-    oidc_rp_scopes_list=["openid"],
+    oidc_rp_scopes_list=["openid", "bsn"],
     oidc_op_jwks_endpoint="http://provider.com/auth/realms/master/protocol/openid-connect/certs",
     oidc_op_authorization_endpoint="http://provider.com/auth/realms/master/protocol/openid-connect/auth",
     oidc_op_token_endpoint="http://provider.com/auth/realms/master/protocol/openid-connect/token",
@@ -29,29 +25,25 @@ default_config = dict(
 
 
 @override_settings(CORS_ALLOW_ALL_ORIGINS=True, IS_HTTPS=True)
-@patch(
-    "digid_eherkenning_oidc_generics.models.OpenIDConnectEHerkenningBewindvoeringConfig.get_solo",
-    return_value=OpenIDConnectEHerkenningBewindvoeringConfig(**default_config),
-)
-class EHerkenningBewindvoeringOIDCTests(TestCase):
+class DigiDOIDCTests(TestCase):
     @classmethod
     def setUpTestData(cls):
         cls.form = FormFactory.create(
-            generate_minimal_setup=True,
-            authentication_backends=["eherkenning_bewindvoering_oidc"],
+            generate_minimal_setup=True, authentication_backends=["digid_oidc"]
         )
 
-    def test_redirect_to_eherkenning_bewindvoering_oidc(self, m_get_solo):
+    @patch(
+        "oidc_generics.models.OpenIDConnectPublicConfig.get_solo",
+        return_value=OpenIDConnectPublicConfig(**default_config),
+    )
+    def test_redirect_to_digid_oidc(self, *m):
         login_url = reverse(
             "authentication:start",
-            kwargs={
-                "slug": self.form.slug,
-                "plugin_id": "eherkenning_bewindvoering_oidc",
-            },
+            kwargs={"slug": self.form.slug, "plugin_id": "digid_oidc"},
         )
 
         form_path = reverse("core:form-detail", kwargs={"slug": self.form.slug})
-        form_url = furl(f"http://testserver{form_path}").set({"_start": "1"}).url
+        form_url = str(furl(f"http://testserver{form_path}").set({"_start": "1"}))
         start_url = furl(login_url).set({"next": form_url})
         response = self.client.get(start_url)
 
@@ -61,10 +53,7 @@ class EHerkenningBewindvoeringOIDCTests(TestCase):
         query_params = parsed.query.params
 
         self.assertEqual(parsed.host, "testserver")
-        self.assertEqual(
-            parsed.path,
-            reverse("eherkenning_bewindvoering_oidc:oidc_authentication_init"),
-        )
+        self.assertEqual(parsed.path, reverse("digid_oidc:oidc_authentication_init"))
 
         parsed = furl(query_params["next"])
         query_params = parsed.query.params
@@ -73,10 +62,7 @@ class EHerkenningBewindvoeringOIDCTests(TestCase):
             parsed.path,
             reverse(
                 "authentication:return",
-                kwargs={
-                    "slug": self.form.slug,
-                    "plugin_id": "eherkenning_bewindvoering_oidc",
-                },
+                kwargs={"slug": self.form.slug, "plugin_id": "digid_oidc"},
             ),
         )
         self.assertEqual(query_params["next"], form_url)
@@ -97,11 +83,11 @@ class EHerkenningBewindvoeringOIDCTests(TestCase):
         self.assertEqual(
             parsed.path, "/auth/realms/master/protocol/openid-connect/auth"
         )
-        self.assertEqual(query_params["scope"], "openid")
+        self.assertEqual(query_params["scope"], "openid bsn")
         self.assertEqual(query_params["client_id"], "testclient")
         self.assertEqual(
             query_params["redirect_uri"],
-            f"http://testserver{reverse('eherkenning_bewindvoering_oidc:oidc_authentication_callback')}",
+            f"http://testserver{reverse('digid_oidc:oidc_authentication_callback')}",
         )
 
         parsed = furl(self.client.session["oidc_login_next"])
@@ -111,23 +97,19 @@ class EHerkenningBewindvoeringOIDCTests(TestCase):
             parsed.path,
             reverse(
                 "authentication:return",
-                kwargs={
-                    "slug": self.form.slug,
-                    "plugin_id": "eherkenning_bewindvoering_oidc",
-                },
+                kwargs={"slug": self.form.slug, "plugin_id": "digid_oidc"},
             ),
         )
         self.assertEqual(query_params["next"], form_url)
 
-    def test_redirect_to_eherkenning_bewindvoering_oidc_internal_server_error(
-        self, m_get_solo
-    ):
+    @patch(
+        "oidc_generics.models.OpenIDConnectPublicConfig.get_solo",
+        return_value=OpenIDConnectPublicConfig(**default_config),
+    )
+    def test_redirect_to_digid_oidc_internal_server_error(self, *m):
         login_url = reverse(
             "authentication:start",
-            kwargs={
-                "slug": self.form.slug,
-                "plugin_id": "eherkenning_bewindvoering_oidc",
-            },
+            kwargs={"slug": self.form.slug, "plugin_id": "digid_oidc"},
         )
 
         form_path = reverse("core:form-detail", kwargs={"slug": self.form.slug})
@@ -149,23 +131,20 @@ class EHerkenningBewindvoeringOIDCTests(TestCase):
 
         self.assertEqual(parsed.host, "testserver")
         self.assertEqual(parsed.path, form_path)
-        self.assertEqual(
-            query_params["of-auth-problem"], "eherkenning_bewindvoering_oidc"
-        )
+        self.assertEqual(query_params["of-auth-problem"], "digid_oidc")
 
-    def test_redirect_to_eherkenning_bewindvoering_oidc_callback_error(
-        self, m_get_solo
-    ):
+    @patch(
+        "oidc_generics.models.OpenIDConnectPublicConfig.get_solo",
+        return_value=OpenIDConnectPublicConfig(**default_config),
+    )
+    def test_redirect_to_digid_oidc_callback_error(self, *m):
         form_path = reverse("core:form-detail", kwargs={"slug": self.form.slug})
         form_url = f"http://testserver{form_path}"
         redirect_form_url = furl(form_url).set({"_start": "1"})
         redirect_url = furl(
             reverse(
                 "authentication:return",
-                kwargs={
-                    "slug": self.form.slug,
-                    "plugin_id": "eherkenning_bewindvoering_oidc",
-                },
+                kwargs={"slug": self.form.slug, "plugin_id": "digid_oidc"},
             )
         ).set({"next": redirect_form_url})
 
@@ -174,12 +153,10 @@ class EHerkenningBewindvoeringOIDCTests(TestCase):
         session.save()
 
         with patch(
-            "openforms.authentication.contrib.digid_eherkenning_oidc.backends.OIDCAuthenticationEHerkenningBewindvoeringBackend.verify_claims",
+            "openforms.authentication.contrib.oidc.backends.OIDCAuthenticationDigiDBackend.verify_claims",
             return_value=False,
         ):
-            response = self.client.get(
-                reverse("eherkenning_bewindvoering_oidc:callback")
-            )
+            response = self.client.get(reverse("digid_oidc:callback"))
 
         self.assertEqual(response.status_code, status.HTTP_302_FOUND)
 
@@ -188,15 +165,16 @@ class EHerkenningBewindvoeringOIDCTests(TestCase):
 
         self.assertEqual(parsed.path, form_path)
         self.assertEqual(query_params["_start"], "1")
-        self.assertEqual(
-            query_params[BACKEND_OUTAGE_RESPONSE_PARAMETER],
-            "eherkenning_bewindvoering_oidc",
-        )
+        self.assertEqual(query_params[BACKEND_OUTAGE_RESPONSE_PARAMETER], "digid_oidc")
 
     @override_settings(CORS_ALLOW_ALL_ORIGINS=False, CORS_ALLOWED_ORIGINS=[])
-    def test_redirect_to_disallowed_domain(self, m_get_solo):
+    @patch(
+        "oidc_generics.models.OpenIDConnectPublicConfig.get_solo",
+        return_value=OpenIDConnectPublicConfig(**default_config),
+    )
+    def test_redirect_to_disallowed_domain(self, *m):
         login_url = reverse(
-            "eherkenning_bewindvoering_oidc:oidc_authentication_init",
+            "digid_oidc:oidc_authentication_init",
         )
 
         form_url = "http://example.com"
@@ -208,25 +186,24 @@ class EHerkenningBewindvoeringOIDCTests(TestCase):
     @override_settings(
         CORS_ALLOW_ALL_ORIGINS=False, CORS_ALLOWED_ORIGINS=["http://example.com"]
     )
-    def test_redirect_to_allowed_domain(self, m_get_solo):
-        m_get_solo.return_value = OpenIDConnectDigiDMachtigenConfig(
+    @patch(
+        "oidc_generics.models.OpenIDConnectPublicConfig.get_solo",
+        return_value=OpenIDConnectPublicConfig(
             enabled=True,
             oidc_rp_client_id="testclient",
             oidc_rp_client_secret="secret",
             oidc_rp_sign_algo="RS256",
-            oidc_rp_scopes_list=["openid"],
+            oidc_rp_scopes_list=["openid", "bsn"],
             oidc_op_jwks_endpoint="http://provider.com/auth/realms/master/protocol/openid-connect/certs",
             oidc_op_authorization_endpoint="http://provider.com/auth/realms/master/protocol/openid-connect/auth",
             oidc_op_token_endpoint="http://provider.com/auth/realms/master/protocol/openid-connect/token",
             oidc_op_user_endpoint="http://provider.com/auth/realms/master/protocol/openid-connect/userinfo",
-        )
-
+        ),
+    )
+    def test_redirect_to_allowed_domain(self, *m):
         login_url = reverse(
             "authentication:start",
-            kwargs={
-                "slug": self.form.slug,
-                "plugin_id": "eherkenning_bewindvoering_oidc",
-            },
+            kwargs={"slug": self.form.slug, "plugin_id": "digid_oidc"},
         )
 
         form_url = "http://example.com"
@@ -239,10 +216,7 @@ class EHerkenningBewindvoeringOIDCTests(TestCase):
         query_params = parsed.query.params
 
         self.assertEqual(parsed.host, "testserver")
-        self.assertEqual(
-            parsed.path,
-            reverse("eherkenning_bewindvoering_oidc:oidc_authentication_init"),
-        )
+        self.assertEqual(parsed.path, reverse("digid_oidc:oidc_authentication_init"))
 
         parsed = furl(query_params["next"])
         query_params = parsed.query.params
@@ -251,10 +225,7 @@ class EHerkenningBewindvoeringOIDCTests(TestCase):
             parsed.path,
             reverse(
                 "authentication:return",
-                kwargs={
-                    "slug": self.form.slug,
-                    "plugin_id": "eherkenning_bewindvoering_oidc",
-                },
+                kwargs={"slug": self.form.slug, "plugin_id": "digid_oidc"},
             ),
         )
         self.assertEqual(query_params["next"], form_url)
@@ -275,33 +246,32 @@ class EHerkenningBewindvoeringOIDCTests(TestCase):
         self.assertEqual(
             parsed.path, "/auth/realms/master/protocol/openid-connect/auth"
         )
-        self.assertEqual(query_params["scope"], "openid")
+        self.assertEqual(query_params["scope"], "openid bsn")
         self.assertEqual(query_params["client_id"], "testclient")
         self.assertEqual(
             query_params["redirect_uri"],
-            f"http://testserver{reverse('eherkenning_bewindvoering_oidc:oidc_authentication_callback')}",
+            f"http://testserver{reverse('digid_oidc:oidc_authentication_callback')}",
         )
 
-    def test_redirect_with_keycloak_identity_provider_hint(self, m_get_solo):
-        m_get_solo.return_value = OpenIDConnectPublicConfig(
+    @patch(
+        "oidc_generics.models.OpenIDConnectPublicConfig.get_solo",
+        return_value=OpenIDConnectPublicConfig(
             enabled=True,
             oidc_rp_client_id="testclient",
             oidc_rp_client_secret="secret",
             oidc_rp_sign_algo="RS256",
-            oidc_rp_scopes_list=["openid"],
+            oidc_rp_scopes_list=["openid", "bsn"],
             oidc_op_jwks_endpoint="http://provider.com/auth/realms/master/protocol/openid-connect/certs",
             oidc_op_authorization_endpoint="http://provider.com/auth/realms/master/protocol/openid-connect/auth",
             oidc_op_token_endpoint="http://provider.com/auth/realms/master/protocol/openid-connect/token",
             oidc_op_user_endpoint="http://provider.com/auth/realms/master/protocol/openid-connect/userinfo",
-            oidc_keycloak_idp_hint="oidc-digid-machtigen",
-        )
-
+            oidc_keycloak_idp_hint="oidc-digid",
+        ),
+    )
+    def test_redirect_with_keycloak_identity_provider_hint(self, *m):
         login_url = reverse(
             "authentication:start",
-            kwargs={
-                "slug": self.form.slug,
-                "plugin_id": "eherkenning_bewindvoering_oidc",
-            },
+            kwargs={"slug": self.form.slug, "plugin_id": "digid_oidc"},
         )
 
         form_path = reverse("core:form-detail", kwargs={"slug": self.form.slug})
@@ -315,10 +285,7 @@ class EHerkenningBewindvoeringOIDCTests(TestCase):
         query_params = parsed.query.params
 
         self.assertEqual(parsed.host, "testserver")
-        self.assertEqual(
-            parsed.path,
-            reverse("eherkenning_bewindvoering_oidc:oidc_authentication_init"),
-        )
+        self.assertEqual(parsed.path, reverse("digid_oidc:oidc_authentication_init"))
 
         parsed = furl(query_params["next"])
         query_params = parsed.query.params
@@ -327,10 +294,7 @@ class EHerkenningBewindvoeringOIDCTests(TestCase):
             parsed.path,
             reverse(
                 "authentication:return",
-                kwargs={
-                    "slug": self.form.slug,
-                    "plugin_id": "eherkenning_bewindvoering_oidc",
-                },
+                kwargs={"slug": self.form.slug, "plugin_id": "digid_oidc"},
             ),
         )
         self.assertEqual(query_params["next"], form_url)
@@ -351,10 +315,10 @@ class EHerkenningBewindvoeringOIDCTests(TestCase):
         self.assertEqual(
             parsed.path, "/auth/realms/master/protocol/openid-connect/auth"
         )
-        self.assertEqual(query_params["scope"], "openid")
+        self.assertEqual(query_params["scope"], "openid bsn")
         self.assertEqual(query_params["client_id"], "testclient")
         self.assertEqual(
             query_params["redirect_uri"],
-            f"http://testserver{reverse('eherkenning_bewindvoering_oidc:oidc_authentication_callback')}",
+            f"http://testserver{reverse('digid_oidc:oidc_authentication_callback')}",
         )
-        self.assertEqual(query_params["kc_idp_hint"], "oidc-digid-machtigen")
+        self.assertEqual(query_params["kc_idp_hint"], "oidc-digid")
