+++ src/openforms/authentication/contrib/oidc/tests/eherkenning/test_admin.py
@@ -3,8 +3,9 @@ from django.urls import reverse
 
 from django_webtest import WebTest
 
-from digid_eherkenning_oidc_generics.models import OpenIDConnectPublicConfig
+from oidc_generics.models import OpenIDConnectEHerkenningConfig
 from openforms.accounts.tests.factories import SuperUserFactory
+from openforms.config.models import GlobalConfiguration
 from openforms.forms.tests.factories import FormFactory
 
 default_config = dict(
@@ -12,7 +13,7 @@ default_config = dict(
     oidc_rp_client_id="testclient",
     oidc_rp_client_secret="secret",
     oidc_rp_sign_algo="RS256",
-    oidc_rp_scopes_list=["openid", "bsn"],
+    oidc_rp_scopes_list=["openid", "kvk"],
     oidc_op_jwks_endpoint="http://provider.com/auth/realms/master/protocol/openid-connect/certs",
     oidc_op_authorization_endpoint="http://provider.com/auth/realms/master/protocol/openid-connect/auth",
     oidc_op_token_endpoint="http://provider.com/auth/realms/master/protocol/openid-connect/token",
@@ -22,24 +23,31 @@ default_config = dict(
 
 
 @override_settings(CORS_ALLOW_ALL_ORIGINS=True, IS_HTTPS=True)
-class DigiDOIDCFormAdminTests(WebTest):
+class eHerkenningOIDCFormAdminTests(WebTest):
     def setUp(self):
         super().setUp()
 
         self.user = SuperUserFactory.create(app=self.app)
         self.app.set_user(self.user)
 
-    def test_digid_oidc_disable_allowed(self):
+        global_config = GlobalConfiguration.get_solo()
+        global_config.enable_react_form = False
+        global_config.save()
+
+        def _cleanup():
+            GlobalConfiguration.get_solo().delete()
+
+        self.addCleanup(_cleanup)
+
+    def test_eherkenning_oidc_disable_allowed(self):
         # Patching `get_solo()` doesn't seem to work when retrieving the change_form
-        config = OpenIDConnectPublicConfig(**default_config)
+        config = OpenIDConnectEHerkenningConfig(**default_config)
         config.save()
 
-        FormFactory.create(authentication_backends=["eherkenning_oidc"])
+        FormFactory.create(authentication_backends=["digid_oidc"])
 
         response = self.app.get(
-            reverse(
-                "admin:digid_eherkenning_oidc_generics_openidconnectpublicconfig_change"
-            )
+            reverse("admin:oidc_generics_openidconnecteherkenningconfig_change")
         )
 
         form = response.form
@@ -47,19 +55,17 @@ class DigiDOIDCFormAdminTests(WebTest):
         response = form.submit()
 
         self.assertEqual(response.status_code, 302)
-        self.assertFalse(OpenIDConnectPublicConfig.get_solo().enabled)
+        self.assertFalse(OpenIDConnectEHerkenningConfig.get_solo().enabled)
 
-    def test_digid_oidc_disable_not_allowed(self):
+    def test_eherkenning_oidc_disable_not_allowed(self):
         # Patching `get_solo()` doesn't seem to work when retrieving the change_form
-        config = OpenIDConnectPublicConfig(**default_config)
+        config = OpenIDConnectEHerkenningConfig(**default_config)
         config.save()
 
-        FormFactory.create(authentication_backends=["digid_oidc"])
+        FormFactory.create(authentication_backends=["eherkenning_oidc"])
 
         response = self.app.get(
-            reverse(
-                "admin:digid_eherkenning_oidc_generics_openidconnectpublicconfig_change"
-            )
+            reverse("admin:oidc_generics_openidconnecteherkenningconfig_change")
         )
 
         form = response.form
@@ -67,4 +73,4 @@ class DigiDOIDCFormAdminTests(WebTest):
         response = form.submit()
 
         self.assertEqual(response.status_code, 200)
-        self.assertTrue(OpenIDConnectPublicConfig.get_solo().enabled)
+        self.assertTrue(OpenIDConnectEHerkenningConfig.get_solo().enabled)
