+++ src/openforms/conf/base.py
@@ -163,7 +163,7 @@ INSTALLED_APPS = [
     "stuf.stuf_zds",
     "mozilla_django_oidc",
     "mozilla_django_oidc_db",
-    "digid_eherkenning_oidc_generics",
+    "oidc_generics",
     "django_filters",
     "csp",
     "cspreports",
@@ -211,7 +211,7 @@ INSTALLED_APPS = [
     "openforms.authentication.contrib.digid_mock.apps.DigidMockApp",
     "openforms.authentication.contrib.digid.apps.DigidApp",
     "openforms.authentication.contrib.eherkenning.apps.EHerkenningApp",
-    "openforms.authentication.contrib.digid_eherkenning_oidc.apps.DigiDEHerkenningOIDCApp",
+    "openforms.authentication.contrib.oidc.apps.OIDCApp",
     "openforms.payments.apps.PaymentsConfig",
     "openforms.payments.contrib.demo.apps.DemoApp",
     "openforms.payments.contrib.ogone.apps.OgoneApp",
